chore: Update Dockerfile and package.json versions and fix the sending message to the topic.
This commit updates the Dockerfile to include copying the "locales" directory to the app directory. It also updates the package.json version from 1.0.10 to 1.0.12. These changes are necessary to support internationalization (i18n) and ensure the latest version is used.
To send message to the topic `replyTo` is used instead of `topMsgId`.